+++ capstone/src/Popup.tsx
@@ -1,0 +1,48 @@
+import * as React from 'react';
+import Button from '@mui/material/Button';
+import { styled } from '@mui/material/styles';
+import Dialog from '@mui/material/Dialog';
+import DialogContent from '@mui/material/DialogContent';
+import DialogActions from '@mui/material/DialogActions';
+
+import ViewPending from './ViewPending';
+
+
+
+export interface DialogTitleProps {
+    id: string;
+    children?: React.ReactNode;
+    onClose: () => void;
+}
+
+export interface formdetails {
+    no: number;
+    buildingPermitNo: string;
+    applicantName: string;
+    projectName: string;
+    open: boolean;
+    handleClose: () => void;
+}
+
+const Popup: React.FC<formdetails> = ({ no, buildingPermitNo, applicantName, projectName, open, handleClose }) => {
+
+    return (
+        <div>
+            <Dialog open={open}>
+                <DialogContent>
+                    <ViewPending
+                        no={no}
+                        buildingPermitNo={buildingPermitNo}
+                        applicantName={applicantName}
+                        projectName={projectName}
+                    />
+                </DialogContent>
+                <DialogActions>
+                    <Button variant='contained' color='error' onClick={handleClose}>Close</Button>
+                </DialogActions>
+            </Dialog>
+
+        </div>
+    );
+};
+export default Popup;
